added module for one pic rendering
